langchain_models/requirements.txt
# LangChain Core
langchain>=0.1.16
langchain-core>=0.1.44

# OpenAI Integration
langchain-openai>=0.1.3
openai>=1.30.1  # Use >=1.0.0 for the new OpenAI SDK structure

# Anthropic Integration
langchain-anthropic>=0.1.3

# Google Gemini (PaLM) Integration
langchain-google-genai>=0.0.7
google-generativeai>=0.3.2

# Hugging Face Integration
langchain-huggingface>=0.0.2
transformers>=4.40.1
huggingface-hub>=0.21.4

# Environment Variable Management
python-dotenv>=1.0.1

# Machine Learning Utilities
numpy>=1.26.4
scikit-learn>=1.4.2
